June1-progress-3
制作了新的UI
进一步优化了哈希表的生成算法
加入了时间测试

+++ Project 4/HashTable.h
@@ -16,7 +16,7 @@ public:
 
 private:
 	std::vector<std::vector<Item>*> m_items;
-	long long m_capacity;
+	int m_capacity;
 };
 
 void fillTable(std::string s, HashTable& table, int SLICELEN);
